requirements.txt
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.7
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
scp==0.14.5
urllib3==1.26.6
